backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pandas==2.2.1
geopandas==0.14.3
shapely==2.0.3
supabase==2.4.1
python-dotenv==1.0.1
python-multipart==0.0.9